PerformStats {COSTsim}	R Documentation

Class "PerformStats"

Description

The PerformStats class stores the performance statistics to evaluate accuracy, bias and precision of different estimators. It is the outcome object after applying the function PerformStats.

Slots

slot 	 class 	 description

desc 	 character 	 object description

species 	 character 	 species description (recall of SL$spp (+ SL$sex))

catchCat 	 character 	 recall of the catch category (discards/landings)

param 	 character 	 recall of the parameter estimated (N, W, maturity, sex-ratio,...)

strataDesc 	 strIni 	 time, space and technical stratification considered

methodDesc 	 character 	 recall of the method (analytical, bootstrap, bayesian)

nSamples 	 numeric 	 number of samples

ageTrue 	 data.frame 	 True age structure (param-at-length)

ageEst 	 data.frame 	 mean estimates of the age structure (param-at-length)

ageAcc 	 list 	 a list with Accuracy statistics at age

ageBias 	 list 	 a list with Bias statistics at age

agePrec 	 list 	 a list with Precision statistics at age

lenTrue 	 data.frame 	 True age structure (param-at-length)

lenEst 	 data.frame 	 mean estimates of the length structure (param-at-length)

lenAcc 	 list 	 a list with Accuracy statistics at length

lenBias 	 list 	 a list with Bias statistics at length

lenPrec 	 list 	 a list with Precision statistics at length

totalNTrue 	 data.frame 	 True age structure (param-at-length)

totalNEst 	 data.frame 	 mean estimates of the length structure (param-at-length)

totalNAcc 	 list 	 a list with Accuracy statistics at length

totalNBias 	 list 	 a list with Bias statistics at length

totalNPrec 	 list 	 a list with Precision statistics at length

totalWTrue 	 data.frame 	 True age structure (param-at-length)

totalWEst 	 data.frame 	 mean estimates of the length structure (param-at-length)

totalWAcc 	 list 	 a list with Accuracy statistics at length

totalWBias 	 list 	 a list with Bias statistics at length

totalWPrec 	 list 	 a list with Precision statistics at length

The accuracy, bias and precision statistics are described in detail in the COST manual

Author(s)

Dorleta García dgarcia@azti.es

Examples


showclass("PerformStats")

[Package COSTsim version 1.0 Index]