RaiseAgeSim {COSTsim}	R Documentation

Estimation of total numbers-at-age from market sampling for simulated data sets

Description

This method is the equivalent of the function RaiseAge for dbeOutputSim class objects and it calculates total numbers-at-age by strata from market sampling for simulated data sets

Usage


RaiseAgeSim(dbeOutputSim,simObj,type="fixed",sex=as.character(NA),...)

Arguments

dbeOutputSim 	 A dbeOutputSim object.

simObj 	 A simDataCons object matching dbeOutputSim specifications.

type 	 Allocation strategy used to establish the age-length data.

sex 	 Sex

... 	 Further arguments

Value

An updated object of class dbeOutputSim. Slots nSamp\$age & nMeas\$age with number of samples and measurements, ageStruc\$estim with numbers-at-age estimates, ageVar with the variance of numbers-at-age .

Author(s)

Dorleta Garcia dgarcia@azti.es

See Also

dbeOutputSim, simDataCons, RaiseAge

[Package COSTsim version 1.0 Index]