RaiseLgthBootSim {COSTsim}	R Documentation

Estimation of total numbers-at-length from market sampling with bootstrap variance for simulated data sets

Description

This method is the equivalent of the function RaiseLgthBoot for dbeOutputSim class objects and it calculates total numbers-at-length by strata from market sampling with bootstrap variance for simulated data sets

Usage


RaiseLgthBootSim(dbeOutputSim,simObj,spp,taxon,sex=as.character(NA),B,...)

Arguments

dbeOutputSim 	 A dbeOutputSim object.

simObj 	 A simDataCons object matching dbeOutputSim specifications.

spp 	 Species, if missing this is set to dbeOutput@species

taxon 	 Taxon, if missing this is set to dbeOutput@species

sex 	 Sex

B 	 Number of bootstrap interations

... 	 Further arguments

Value

An updated object of class dbeOutputSim. Slots nSamp\$len & nMeas\$len with number of samples and measurements, methodDesc with "bootstrap", totalW\$estim with total weight, lenStruc\$rep & totalN\$rep with bootstrap replicates for numbers-at-length and total numbers, iter=0 is assigned the estimates from the original data, lenStruc\$estim & totalN\$estim with the mean of the bootstrap replicates, lenVar & totalNvar with the variance of the bootstrap replicates

Author(s)

Dorleta Garcia <dgarcia@azti.es>

See Also

dbeOutputSim, simDataCons, RaiseLgthBoot

[Package COSTsim version 1.0 Index]