RaiseLgthSim {COSTsim}	R Documentation

Estimation of total numbers-at-length from market sampling for simulated data sets

Description

This method is the equivalent of the function RaiseLgth for dbeOutputSim class objects and it calculates total numbers-at-length by strata from market sampling for simulated data sets

Usage


RaiseLgthSim(dbeOutputSim,simObj,spp,taxon,sex=as.character(NA),...)

Arguments

dbeOutputSim 	 A dbeOutputSim object.

simObj 	 A simDataCons object matching 'dbeOutputSim' specifications.

spp 	 Species, if missing this is set to dbeOutput@species

taxon 	 Taxon, if missing this is set to dbeOutput@species

sex 	 Sex

... 	 Further arguments

Value

An updated object of class dbeOutputSim. Slots nSamp\$len & nMeas\$len with number of samples and measurements, totalW\$estim with total weight, lenStruc\$estim with numbers-at-length estimates, lenVar with the variance of numbers-at-length.

Author(s)

Dorleta Garcia <dgarcia@azti.es>

See Also

dbeOutputSim, simDataCons, RaiseLgth

[Package COSTsim version 1.0 Index]