alkLgthRecSim {COSTsim}	R Documentation

Method for managing gaps in age-length keys for simulated data sets

Description

This function is the counterpart of the alkLgthRec function for simDataCons objects. It provides various methods to solve alk gaps problems. Input object is updated according to chosen method(s) (grouped/recoded length classes, addition of 'virtual' individuals,...)

Usage


alkLgthRecSim(object,type="stepIncr",value,preview=FALSE,postview=TRUE,update=FALSE,...)

Arguments

object 	 A simDataCons object with simulated data sets

type 	 Character for chosen method. Values are :
"stepIncr"

Default parameter. Length class step is increased to specified value parameter (default value=10)

"fillMiss"

All gaps (with size <= value) are filled out with the sum of surrounding recorded classes (default value=1)

"sFillMiss"

The 'value' empty classe(s) prior to first recorded length class is filled out with the latter (default value=1)

"lFillMiss"

The 'value' empty classe(s) following last recorded length class is filled out with the latter (default value=1)

value 	 Numerical parameter for chosen method (see 'type').

preview 	 Logical. If TRUE, original age length key is displayed.

postview 	 Logical. If TRUE, new age length key is displayed.

update 	 Logical. If TRUE, 'csDataCons' object is updated in accordance with chosen method, and then returned. If FALSE, descriptive elements about updated alk are returned (see 'values'), but input object remains unchanged.

... 	 Further arguments, and particularly a start numerical parameter specifying the first considered length class when recoding (only useful for 'type="stepIncr"'). Default value is the minimum aged length class in ca table.

Value

If update=FALSE, returned elements within @samples of the simDataCons object are : $alk is the raw resulting age-length key, $propMiss are short statistics about gaps (see 'propMissLgthCons' method), $lgthCls is a description of length classes recoding for 'stepIncr', 'sExtrGrp' and 'lExtrGrp' methods and $addIndTab is a description of added virtual individuals for other methods.

Author(s)

Dorleta garcia dgarcia@azti.es

See Also

alkLgthRec

[Package COSTsim version 1.0 Index]