costDataCons {COSTsim}	R Documentation

Class "costDataCons"

Description

The costDataCons-class is equivalent in structure to costData but stores consolidated forms of the simulated data set

Objects from the Class

The creator function costDataCons can be called to create objects from this class.

Slots

slot 	 class 	 description

desc 	 character 	 object description

ce 	 ceDataCons 	 consolidated commercial fisheries effort data

cl 	 clDataCons 	 consolidated commercial fisheries landings data

cs 	 csDataCons 	 consolidated commercial fisheries sampling data

Author(s)

Dorleta García dgarcia@azti.es

Examples


showclass("costDataCons")

[Package COSTsim version 1.0 Index]