costDataVal {COSTsim}	R Documentation

Class "costDataVal"

Description

The costDataVal-class is equivalent in structure to costData but stores validated forms of the simulated data set

Objects from the Class

The creator function "costDataVal" can be called to create objects from this class.

Slots

slot 	 class 	 description

desc 	 character 	 object description

ce 	 ceDataVal 	 validated commercial fisheries effort data

cl 	 clDataVal 	 validated commercial fisheries landings data

cs 	 csDataVal 	 validated commercial fisheries sampling data

Author(s)

Dorleta Garcia dgarcia@azti.es

Examples


showclass("costDataVal")

[Package COSTsim version 1.0 Index]