dbeCalcSim {COSTsim}	R Documentation

CI and CV calculation

Description

Method for calculating coefficients of variation or confidence intervals from dbeOutputSim object estimates. This function is equivalent to dbeCalc for dbeOutput but with the update argument always set equal to TRUE

Usage


dbeCalcSim(object,type="CI",vrbl="l",probs=c(0.025,0.975),replicates=FALSE,...)

Arguments

object 	 A dbeOutputSim object.

type 	 Character. "CI" for confidence interval calculation, or "CV" for coefficients of variation.

vrbl 	 Character specifying 'dbeOutput' estimates on which calculation is applied : "l" for length structure, "a" for age structure, "n" for total number estimates, "w" for total weight estimates.

probs 	 Numeric vector of probabilities with values in [0,1]. Defines CI bounds (relevant only if type="CI"). See quantile.

replicates 	 Logical. If TRUE, calculation is made from @...$rep elements ; if FALSE, @...$estim and @...Var 'dbeOutput' data are used.

... 	 Further arguments used as 'quantile method input parameter (if type="CI" and besides probs parameter).

Details

If calculation is made from replicates (see replicates parameter), confidence interval is estimated using quantile function with probs and ... parameters. If calculation is made from estimates, normal distribution of total estimates is assumed to compute confidence intervals. Possible resulting negative bounds are automatically replaced by 0 in output object.

Author(s)

Dorleta Garcia dgarcia@azti.es

See Also

dbeOutputSim, dbeCalc quantile

[Package COSTsim version 1.0 Index]