dbeOutputSim {COSTsim}	R Documentation

Class "dbeOutputSim"

Description

Outcome object from applying COSTdbe methods to simData class objects. It has the same main estructure as dbeOutput class objects but the dataframes generated by the estimation methods include a variable sample which identifies the simulated data set.

Objects from the Class

The creator function dbeOutputSim can be called to create objects from this class.

Slots

slot 	 desc 	 elements 	 class 	 description

desc 	 	 	 character 	 Descriptive slot

species 	 	 	 character 	 Species

catchCat 	 	 	 character 	 Catch category (eg "LAN", "DIS" or "all")

param 	 	 	 character 	 Parameter estimated (eg "maturity","sex-ratio",...)

strataDesc 	 	 	 strIni 	 Stratification considered

methodDesc 	 	 	 character 	 Used method (eg "analytical","bootstrap" or "Bayesian")

All the following slots contain numerics or dataframes that will be generated by COSTdbe methods :

nSamp 	 	 	 list 	 Number of samples

len 	 data.frame 	 Number of length samples.

age 	 data.frame 	 Number of age samples.

nMes 	 	 	 list 	 Number of individual measured

len 	 data.frame 	 Number of individual measured for length data.

age 	 data.frame 	 Number of individual measured for age data.

lenStruc 	 	 	 list 	 Estimates of the length structure

estim 	 data.frame 	 Final estimates (length field added).

rep 	 data.frame 	 Resampling replicates (length and iter fields added).

lenVar 	 	 	 data.frame 	 Estimates of the variance of 'lenStruc'

lenNum 	 	 	 list 	 Precision estimates for length structure

ci 	 data.frame 	 Confidence intervals.

cv 	 data.frame 	 Coefficients of variation.

DCRcvIndicator 	 numeric 	 Weighted global cv.

ageStruc 	 	 	 list 	 Estimates of the age structure

estim 	 data.frame 	 Final estimates (age field added).

rep 	 data.frame 	 Resampling replicates (age and iter fields added).

ageVar 	 	 	 data.frame 	 Estimates of the variance of 'ageStruc'

ageNum 	 	 	 list 	 Precision estimates for age structure

ci 	 data.frame 	 Confidence intervals.

cv 	 data.frame 	 Coefficients of variation.

DCRcvIndicator 	 numeric 	 Weighted global cv.

totalN 	 	 	 list 	 Estimates of the total number of the parameters

estim 	 data.frame 	 Final estimates.

rep 	 data.frame 	 Resampling replicates (iter field added).

totalNvar 	 	 	 data.frame 	 Estimates of the variance of 'totalN'

totalNnum 	 	 	 list 	 Precision estimates for total numbers

ci 	 data.frame 	 Confidence intervals.

cv 	 data.frame 	 Coefficients of variation.

DCRcvIndicator 	 numeric 	 Weighted global cv.

totalW 	 	 	 list 	 Estimates of the total weight of the parameters

estim 	 data.frame 	 Final estimates.

rep 	 data.frame 	 Resampling replicates (iter field added).

totalWvar 	 	 	 data.frame 	 Estimates of the variance of 'totalW'

totalWnum 	 	 	 list 	 Precision estimates for total weights

ci 	 data.frame 	 Confidence intervals.

cv 	 data.frame 	 Coefficients of variation.

DCRcvIndicator 	 numeric 	 Weighted global cv.

Author(s)

Dorleta Garcia dgarcia@azti.es

See Also

dbeOutput

Examples


showClass("dbeOutputSim")

[Package COSTsim version 1.0 Index]