fillGaps {COSTsim}	R Documentation

Method for completing gaps in simulated data sets

Description

This method completes dbeSimObj class objects with zeros in non sampled age and length dimensions in order to have objects of the same dimension that can be compared

Usage


fillGaps(dbeSimObj, ageMin, ageMax, lenMin, lenMax)

Arguments

dbeSimObj 	 A dbeOutputSim object

ageMin 	 Numeric indicating the minimum age sampled

ageMax 	 Numeric indicating the maximum age sampled

lenMin 	 Numeric indicating the minimum length sampled

lenMax 	 Numeric indicating the maximum length sampled

Value

It returns the same object but with the age and length dimensions completed with zeros

Author(s)

Dorleta Garcia dgarcia@azti.es

[Package COSTsim version 1.0 Index]