simData {COSTsim}	R Documentation

Class "simData"

Description

The simData class stores various costData class objects simulated according to the Bayesian model in library COSTmbe

Objects from the Class

The creator function simData can be called to create objects from this class.

Slots

slot 	 class 	 description

desc 	 character 	 object description

species 	 character 	 species description. Recall of SL$spp and SL$sex

samples 	 list 	 each element in the list is a costData object representing a simulated data set

initial.fit 	 list

setup.args 	 list 	 set up parameters

burnin 	 numeric

nmcmc 	 numeric

l.int 	 numeric

Int 	 list

Slp 	 list

landings 	 numeric

nHaul 	 integer

nseas 	 numeric

Author(s)

Dorleta García dgarcia@azti.es

Examples


showclass("simData")

[Package COSTsim version 1.0 Index]