simSamples {COSTsim}	R Documentation

Simulation of data sets

Description

This function simulates data sets for the evaluation and comparison of the estimation methods

Usage


simSamples(obj, ...)

Arguments

obj 	 a 'simData' object

... 	 Further arguments

Value

It returns a simData object with the sample slots containing the simulated data sets

Author(s)

Dorleta Garcia dgarcia@azti.es

See Also

simData

[Package COSTsim version 1.0 Index]