totVolumeSim {COSTsim}	R Documentation

Estimation of total volume of discards or/and landings (weight, number or number-at-length) of simulated data sets

Description

This function is the equivalent to totVolume for 'simDataCons' class objects. It estimates total volume of discards or/and landings (weight, number or number-at-length) based on various raising methods for simulated data sets.

Usage


totVolume(dbeOutputSim,simObject,...)

Arguments

dbeOutputSim 	 A dbeOutputSim object. All necessary information for calculation process are taken in the first slots (species, catch category,...). See dbeObject method for object initialization.

simObject 	 A simDataCons object matching 'dbeOutputSim' specifications.

... 	 Further arguments such as:
type

Specification of the raising method : "trip" (default value) for raising by trip, "fo" for raising by fishing operations, "fd" for raising by fishing days,"landings" for ratio-to-total landings raising method, and "time" for ratio-to-fishing duration raising method.

val

Estimated parameter. To be chosen between "weight" (default value), "number" and "nAtLength".

sampPar

logical specifying if given species is considered to be automatically sampled during the sampling process (default value is TRUE).

landSpp

character vector describing the species considered in the 'volume of landings' variable if chosen raising method is ratio-to-landings (see 'clObject' description).

Value

An updated object of class dbeOutputSim.

Author(s)

Dorleta Garcia <dgarcia@azti.es>

References

Vigneau, J. (2006) Raising procedures for discards : Sampling theory (Toward agreed methodologies for calculating precision in the discard programmes). Working document in support of PGCCDBS (Rostock, 2006).

See Also

totVolume

[Package COSTsim version 1.0 Index]